DetailedFragment implemented

+++ app/src/main/java/ua/mrrobot1413/movies/data/network/model/MoviesResponse.kt
@@ -12,7 +12,7 @@ data class MoviesResponse(
 
 data class Movie(
     @SerializedName("id")
-    val id: BigInteger,
+    val id: Int,
     @SerializedName("title")
     val title: String,
     @SerializedName("adult")
